Move GCP system tests to providers.google.cloud
