+++ tests/providers/google/cloud/operators/test_gcs_to_gcs_system_helper.py
