+++ tests/providers/google/cloud/operators/test_gcs_to_sftp_system.py
@@ -18,7 +18,7 @@
 # under the License.
 """System tests for Google Cloud Build operators"""
 
-from tests.operators.test_gcs_to_sftp_system_helper import GcsToSFTPTestHelper
+from tests.providers.google.cloud.operators.test_gcs_to_sftp_system_helper import GcsToSFTPTestHelper
 from tests.providers.google.cloud.utils.gcp_authenticator import GCP_GCS_KEY
 from tests.test_utils.gcp_system_helpers import CLOUD_DAG_FOLDER, provide_gcp_context, skip_gcp_system
 from tests.test_utils.system_tests_class import SystemTest
